requirements.txt
click
google-cloud-asset
google-cloud-org-policy
google-cloud-compute